CodebuSTer1.0/requirements.txt
customtkinter
python-nmap
requests
geoip2
python-whois
Pillow
beautifulsoup4
dnspython 
PyPDF2
python-docx
ExifRead 
